FM2-185: Moving back CIEL constants to impl class.
