+++ api/src/main/java/org/openmrs/module/fhir2/api/translators/impl/ImmunizationTranslatorImpl.java
@@ -49,6 +49,23 @@ import org.springframework.stereotype.Component;
 @Setter(AccessLevel.PACKAGE)
 public class ImmunizationTranslatorImpl implements ImmunizationTranslator {
 	
+	public static final String immunizationGroupingConcept = "CIEL:1421";
+	
+	public static final String[] immunizationConcepts = { "CIEL:984", "CIEL:1410", "CIEL:1418", "CIEL:1419", "CIEL:1420",
+	        "CIEL:165907" };
+	
+	public static final String ciel984 = immunizationConcepts[0];
+	
+	public static final String ciel1410 = immunizationConcepts[1];
+	
+	public static final String ciel1418 = immunizationConcepts[2];
+	
+	public static final String ciel1419 = immunizationConcepts[3];
+	
+	public static final String ciel1420 = immunizationConcepts[4];
+	
+	public static final String ciel165907 = immunizationConcepts[5];
+	
 	private ConceptService conceptService;
 	
 	@Autowired
